pyproject.toml
[tool.poetry]
name = "fileshare"
version = "0.1.0"
description = ""
authors = ["ell <ell@ell.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
django-environ = "^0.11.2"
redis = "^5.0.1"
django-redis = "^5.4.0"
psycopg2-binary = "^2.9.9"
django-taggit = "^5.0.1"
shortuuid = "^1.0.11"
django-storages = "^1.14.2"
boto3 = "^1.34.2"
whitenoise = "^6.6.0"
django-browser-reload = "^1.12.1"
celery = "^5.3.6"
ffmpeg-python = "^0.2.0"
Pillow = "^10.1.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
black = "^23.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
